Refactor iconCss
