+++ src/containers/MyNdla/Folders/FolderCreateModal.tsx
@@ -14,9 +14,8 @@ import { breakpoints, mq } from "@ndla/core";
 import { Plus } from "@ndla/icons/action";
 import { Modal, ModalBody, ModalCloseButton, ModalContent, ModalHeader, ModalTitle, ModalTrigger } from "@ndla/modal";
 import FolderForm, { FolderFormValues } from "./FolderForm";
-import { iconCss } from "./FoldersPage";
 import { GQLFolder } from "../../../graphqlTypes";
-import { buttonCss } from "../components/toolbarStyles";
+import { buttonCss, iconCss } from "../components/toolbarStyles";
 import { useAddFolderMutation, useFolders } from "../folderMutations";
 
 const AddButton = styled(ButtonV2)`
