+++ src/containers/MyNdla/Folders/FolderDeleteModal.tsx
@@ -11,9 +11,8 @@ import { useTranslation } from "react-i18next";
 import { ButtonV2 } from "@ndla/button";
 import { TrashCanOutline } from "@ndla/icons/action";
 import { Modal, ModalTrigger } from "@ndla/modal";
-import { iconCss } from "./FoldersPage";
 import DeleteModalContent from "../components/DeleteModalContent";
-import { buttonCss } from "../components/toolbarStyles";
+import { buttonCss, iconCss } from "../components/toolbarStyles";
 
 interface Props {
   onDelete: () => void;
